pebble.png
added facts: Rock or stone is a natural substance, a solid aggregate of one or more minerals or mineraloids. For example, granite, a common rock, is a combination of the minerals quartz, feldspar and biotite. The Earth's outer solid layer, the lithosphere, is made of rock.

Rock has been used by humankind throughout history. The minerals and metals in rocks have been essential to human civilization.[1]

Three major groups of rocks are defined: igneous, sedimentary, and metamorphic. The scientific study of rocks is called petrology, which is an essential component of geology.